+++ app/fdroid/release/output.json
@@ -1,0 +1,1 @@
+[{"outputType":{"type":"APK"},"apkInfo":{"type":"MAIN","splits":[],"versionCode":10,"versionName":"1.0.9-FDroid","enabled":true,"outputFile":"app-fdroid-release.apk","fullName":"fdroidRelease","baseName":"fdroid-release"},"path":"app-fdroid-release.apk","properties":{}}]
