+++ app/playStore/release/output.json
@@ -1,0 +1,1 @@
+[{"outputType":{"type":"APK"},"apkInfo":{"type":"MAIN","splits":[],"versionCode":10,"versionName":"1.0.9-PlayStore","enabled":true,"outputFile":"app-playStore-release.apk","fullName":"playStoreRelease","baseName":"playStore-release"},"path":"app-playStore-release.apk","properties":{}}]
