+++ app/src/main/java/org/fossasia/susi/ai/chat/ParseSusiResponseHelper.kt
@@ -3,9 +3,8 @@ package org.fossasia.susi.ai.chat
 import android.util.Patterns
 import io.realm.RealmList
 import org.fossasia.susi.ai.data.model.MapData
-import org.fossasia.susi.ai.data.model.TableDatas
+import org.fossasia.susi.ai.data.model.TableItem
 import org.fossasia.susi.ai.helper.Constant
-import org.fossasia.susi.ai.rest.responses.susi.Answer
 import org.fossasia.susi.ai.rest.responses.susi.Datum
 import org.fossasia.susi.ai.rest.responses.susi.SusiResponse
 import timber.log.Timber
@@ -26,7 +25,7 @@ class ParseSusiResponseHelper {
     var webSearch = ""
     var stop = "Stopped"
     var isHavingLink = false
-    var tableData: TableDatas? = null
+    var tableData: TableItem? = null
 
     fun parseSusiResponse(susiResponse: SusiResponse, i: Int, error: String) {
 
@@ -99,7 +98,7 @@ class ParseSusiResponseHelper {
                             listTableData.add(it[i].toString())
                         }
                     }
-                    tableData = TableDatas(listColVal, listTableData)
+                    tableData = TableItem(listColVal, listTableData)
                 }
             } catch (e: Exception) {
                 Timber.e(e)
